refactor: improve example structure
