+++ example/components/TravelsModal/types.ts
@@ -1,0 +1,7 @@
+import { Travel } from "../../shared/types";
+
+export interface TravelsModalProps {
+  isModalOpen: boolean;
+  selectedTravels: Travel[];
+  handleCloseModal: () => void;
+}
