+++ example/components/ui/types.ts
@@ -1,0 +1,3 @@
+export interface AspectRatioProps {
+  ratio: number;
+}
